+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Added
 - Rename ``--cassette-*`` options -> ``--report=vcr/har`` with format-specific paths.
 - Replace ``shrink`` in ``--hypothesis-phases`` with a separate ``--no-shrink`` option.
 - Simplified filtering options help message.
+- Display the number of selected operations in CLI.
 
 **Fixed**